回调抽离
